+++ include/pinocchio/algorithm/rnea-second-order-derivatives.hpp
@@ -162,6 +162,12 @@ namespace pinocchio
     const Eigen::MatrixBase<TangentVectorType1> & v,
     const Eigen::MatrixBase<TangentVectorType2> & a)
   {
+
+    PINOCCHIO_CHECK_INPUT_ARGUMENT(
+      data.allocation == DataAllocationOption::ALL,
+      "data was created with DataAllocationOption::NO_TENSORS: tensors are not "
+      "allocated");
+
     (data.d2tau_dqdq).setZero();
     (data.d2tau_dvdv).setZero();
     (data.d2tau_dqdv).setZero();